Debug attempt and organizing
